+++ utxo-selection-benchmark/Cargo.toml
@@ -23,3 +23,4 @@ serde_yaml = "0.9.13"
 clap = { version = "3.1", features = ["derive"] }
 cardano-multiplatform-lib = "3.1.1"
 pallas-addresses = "0.16.0"
+reqwest = { version = "0.11", features = ["json"] }
