+++ utxo-selection-benchmark/src/lib.rs
@@ -1,19 +1,5 @@
 pub mod bench;
+pub mod generation_utils;
 pub mod mapper;
 pub mod tx_event;
 pub mod utils;
-
-pub fn add(left: usize, right: usize) -> usize {
-    left + right
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn it_works() {
-        let result = add(2, 2);
-        assert_eq!(result, 4);
-    }
-}
